Fix UI Bugs
